backend/requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
pyserial==3.5
Flask-RESTful==0.3.9
Flask-JSON
Flask-SQLAlchemy==2.5.1